Struct google_admin1_directory::Directory

pub struct Directory<C, A> {
    // some fields omitted
}

Central instance to access all Directory related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_admin1_directory as admin1_directory;
use admin1_directory::Channel;
use admin1_directory::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use admin1_directory::Directory;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Directory::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Channel::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.users().watch(req)
             .view_type("labore")
             .sort_order("sea")
             .show_deleted("nonumy")
             .query("dolores")
             .projection("gubergren")
             .page_token("sadipscing")
             .order_by("aliquyam")
             .max_results(-66)
             .event("no")
             .domain("justo")
             .customer("justo")
             .custom_field_mask("et")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Methods

impl<'a, C, A> Directory<C, A> where C: BorrowMut<Client>, A: GetToken

fn new(client: C, authenticator: A) -> Directory<C, A>

fn asps(&'a self) -> AspMethods<'a, C, A>

fn channels(&'a self) -> ChannelMethods<'a, C, A>

fn chromeosdevices(&'a self) -> ChromeosdeviceMethods<'a, C, A>

fn customers(&'a self) -> CustomerMethods<'a, C, A>

fn domain_aliases(&'a self) -> DomainAliaseMethods<'a, C, A>

fn domains(&'a self) -> DomainMethods<'a, C, A>

fn groups(&'a self) -> GroupMethods<'a, C, A>

fn members(&'a self) -> MemberMethods<'a, C, A>

fn mobiledevices(&'a self) -> MobiledeviceMethods<'a, C, A>

fn notifications(&'a self) -> NotificationMethods<'a, C, A>

fn orgunits(&'a self) -> OrgunitMethods<'a, C, A>

fn privileges(&'a self) -> PrivilegeMethods<'a, C, A>

fn resources(&'a self) -> ResourceMethods<'a, C, A>

fn role_assignments(&'a self) -> RoleAssignmentMethods<'a, C, A>

fn roles(&'a self) -> RoleMethods<'a, C, A>

fn schemas(&'a self) -> SchemaMethods<'a, C, A>

fn tokens(&'a self) -> TokenMethods<'a, C, A>

fn users(&'a self) -> UserMethods<'a, C, A>

fn verification_codes(&'a self) -> VerificationCodeMethods<'a, C, A>

fn user_agent(&mut self, agent_name: String) -> String

Set the user-agent header field to use in all requests to the server. It defaults to google-api-rust-client/0.1.15.

Returns the previously set user-agent.