Struct google_admin1_directory::Group

pub struct Group {
    pub non_editable_aliases: Option<Vec<String>>,
    pub kind: Option<String>,
    pub description: Option<String>,
    pub admin_created: Option<bool>,
    pub direct_members_count: Option<i64>,
    pub email: Option<String>,
    pub etag: Option<String>,
    pub aliases: Option<Vec<String>>,
    pub id: Option<String>,
    pub name: Option<String>,
}

JSON template for Group resource in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get groups (response)

aliases insert groups (none)

delete groups (none)

aliases delete groups (none)

patch groups (request|response)

list groups (none)

aliases list groups (none)

update groups (request|response)

insert groups (request|response)

Fields

non_editable_aliases

List of non editable aliases (Read-only)

kind

Kind of resource this is.

description

Description of the group

admin_created

Is the group created by admin (Read-only) *

direct_members_count

Group direct members count

email

Email of Group

etag

ETag of the resource.

aliases

List of aliases (Read-only)

id

Unique identifier of Group (Read-only)

name

Group name