Struct google_admin1_directory::Notification

pub struct Notification {
    pub body: Option<String>,
    pub kind: Option<String>,
    pub etag: Option<String>,
    pub send_time: Option<String>,
    pub notification_id: Option<String>,
    pub from_address: Option<String>,
    pub is_unread: Option<bool>,
    pub subject: Option<String>,
}

Template for a notification resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete notifications (none)

update notifications (request|response)

patch notifications (request|response)

get notifications (response)

list notifications (none)

Fields

body

Body of the notification (Read-only)

kind

The type of the resource.

etag

ETag of the resource.

send_time

Time at which notification was sent (Read-only)

notification_id

no description provided

from_address

Address from which the notification is received (Read-only)

is_unread

Boolean indicating whether the notification is unread or not.

subject

Subject of the notification (Read-only)