Struct google_admin1_directory::OrgunitMethods

pub struct OrgunitMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on orgunit resources. It is not used directly, but through the Directory hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_admin1_directory as admin1_directory;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use admin1_directory::Directory;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Directory::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.orgunits();

Methods

impl<'a, C, A> OrgunitMethods<'a, C, A>

fn insert(&self, request: OrgUnit, customer_id: &str) -> OrgunitInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Add Organization Unit

Arguments

request - No description provided.

customerId - Immutable id of the Google Apps account

fn list(&self, customer_id: &str) -> OrgunitListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieve all Organization Units

Arguments

customerId - Immutable id of the Google Apps account

fn get(&self, customer_id: &str, org_unit_path: &Vec<String>) -> OrgunitGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieve Organization Unit

Arguments

customerId - Immutable id of the Google Apps account

orgUnitPath - Full path of the organization unit or its Id

fn update(&self, request: OrgUnit, customer_id: &str, org_unit_path: &Vec<String>) -> OrgunitUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Update Organization Unit

Arguments

request - No description provided.

customerId - Immutable id of the Google Apps account

orgUnitPath - Full path of the organization unit or its Id

fn delete(&self, customer_id: &str, org_unit_path: &Vec<String>) -> OrgunitDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Remove Organization Unit

Arguments

customerId - Immutable id of the Google Apps account

orgUnitPath - Full path of the organization unit or its Id

fn patch(&self, request: OrgUnit, customer_id: &str, org_unit_path: &Vec<String>) -> OrgunitPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Update Organization Unit. This method supports patch semantics.

Arguments

request - No description provided.

customerId - Immutable id of the Google Apps account

orgUnitPath - Full path of the organization unit or its Id