Struct google-admin1_directory::Token

pub struct Token {
    pub scopes: Vec<String>,
    pub kind: String,
    pub client_id: String,
    pub native_app: bool,
    pub display_text: String,
    pub etag: String,
    pub anonymous: bool,
    pub user_key: String,
}

JSON template for token resource in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list tokens (none)

delete tokens (none)

get tokens (response)

Fields

scopes

A list of authorization scopes the application is granted.

kind

The type of the API resource. This is always admin#directory#token.

client_id

The Client ID of the application the token is issued to.

native_app

Whether the token is issued to an installed application. The value is true if the application is installed to a desktop or mobile device.

display_text

The displayable name of the application the token is issued to.

etag

ETag of the resource.

anonymous

Whether the application is registered with Google. The value is true if the application has an anonymous Client ID.

user_key

The unique ID of the user that issued the token.