Struct google_admin1_directory::TokenMethods

pub struct TokenMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on token resources. It is not used directly, but through the Directory hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_admin1_directory as admin1_directory;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use admin1_directory::Directory;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Directory::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)` and `list(...)`
// to build up your call.
let rb = hub.tokens();

Methods

impl<'a, C, A> TokenMethods<'a, C, A>

fn get(&self, user_key: &str, client_id: &str) -> TokenGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get information about an access token issued by a user.

Arguments

userKey - Identifies the user in the API request. The value can be the user's primary email address, alias email address, or unique user ID.

clientId - The Client ID of the application the token is issued to.

fn list(&self, user_key: &str) -> TokenListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the set of tokens specified user has issued to 3rd party applications.

Arguments

userKey - Identifies the user in the API request. The value can be the user's primary email address, alias email address, or unique user ID.

fn delete(&self, user_key: &str, client_id: &str) -> TokenDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete all access tokens issued by a user for an application.

Arguments

userKey - Identifies the user in the API request. The value can be the user's primary email address, alias email address, or unique user ID.

clientId - The Client ID of the application the token is issued to.