Struct google-admin1_directory::UserAddress

pub struct UserAddress {
    pub source_is_structured: bool,
    pub custom_type: String,
    pub country_code: String,
    pub locality: String,
    pub country: String,
    pub region: String,
    pub po_box: String,
    pub formatted: String,
    pub primary: bool,
    pub street_address: String,
    pub postal_code: String,
    pub extended_address: String,
    pub type_: String,
}

JSON template for address.

This type is not used in any activity, and only used as part of another schema.

Fields

source_is_structured

User supplied address was structured. Structured addresses are NOT supported at this time. You might be able to write structured addresses, but any values will eventually be clobbered.

custom_type

Custom type.

country_code

Country code.

locality

Locality.

country

Country.

region

Region.

po_box

Other parts of address.

formatted

Formatted address.

primary

If this is user's primary address. Only one entry could be marked as primary.

street_address

Street.

postal_code

Postal code.

extended_address

Extended Address.

type_

Each entry can have a type which indicates standard values of that entry. For example address could be of home, work etc. In addition to the standard type, an entry can have a custom type and can take any value. Such type should have the CUSTOM value as type and also have a customType value.