Crate google-admin1_directory

This documentation was generated from directory crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

asps (delete, get and list)

channels (stop)

chromeosdevices (get, list, patch and update)

groups (aliases.delete, aliases.insert, aliases.list, delete, get, insert, list, patch and update)

members (delete, get, insert, list, patch and update)

mobiledevices (action, delete, get and list)

notifications (delete, get, list, patch and update)

orgunits (delete, get, insert, list, patch and update)

schemas (delete, get, insert, list, patch and update)

tokens (delete, get and list)

users (aliases.delete, aliases.insert, aliases.list, aliases.watch, delete, get, insert, list, make admin, patch, photos.delete, photos.get, photos.patch, photos.update, undelete, update and watch)

verification codes (generate, invalidate and list)

Subscription supported by ...

list users

aliases.list users

watch users

aliases.watch users

aliases.list groups

Everything else about the directory v1_directory API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.users().photos_patch(...).doit()
let r = hub.users().aliases_delete(...).doit()
let r = hub.users().undelete(...).doit()
let r = hub.users().photos_get(...).doit()
let r = hub.users().update(...).doit()
let r = hub.users().aliases_watch(...).doit()
let r = hub.users().insert(...).doit()
let r = hub.users().photos_delete(...).doit()
let r = hub.users().patch(...).doit()
let r = hub.users().photos_update(...).doit()
let r = hub.users().watch(...).doit()
let r = hub.users().get(...).doit()
let r = hub.users().aliases_insert(...).doit()
let r = hub.users().make_admin(...).doit()
let r = hub.users().aliases_list(...).doit()
let r = hub.users().list(...).doit()
let r = hub.users().delete(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-admin1_directory = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-admin1_directory" as admin1_directory;
use admin1_directory::Channel;
use admin1_directory::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use admin1_directory::Directory;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Directory::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: Channel = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.users().watch(&req)
             .view_type("labore")
             .sort_order("sea")
             .show_deleted("nonumy")
             .query("dolores")
             .projection("gubergren")
             .page_token("sadipscing")
             .order_by("aliquyam")
             .max_results(-66)
             .event("no")
             .domain("justo")
             .customer("justo")
             .custom_field_mask("et")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Uploads and Downlods

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Structs

Alias
JSON template for Alias object in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

aliases insert users (request|response)

aliases insert groups (request|response)

Aliases
JSON response template to list aliases in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

aliases list users (response)

aliases list groups (response)

Asp
The template that returns individual ASP (Access Code) data.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list asps (none)

delete asps (none)

get asps (response)

AspDeleteMethodBuilder
Delete an ASP issued by a user.

AspGetMethodBuilder
Get information about an ASP issued by a user.

AspListMethodBuilder
List the ASPs issued by a user.

AspMethodsBuilder
A builder providing access to all methods supported on asp resources. It is not used directly, but through the Directory hub.

Asps
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list asps (response)

Channel
An notification channel used to watch for resource changes.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

watch users (request|response)

stop channels (request)

aliases watch users (request|response)

ChannelMethodsBuilder
A builder providing access to all methods supported on channel resources. It is not used directly, but through the Directory hub.

ChannelStopMethodBuilder
Stop watching resources through this channel

ChromeOsDevice
JSON template for Chrome Os Device resource in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get chromeosdevices (response)

update chromeosdevices (request|response)

patch chromeosdevices (request|response)

ChromeOsDeviceActiveTimeRanges
List of active time ranges (Read-only)

This type is not used in any activity, and only used as part of another schema.

ChromeOsDeviceRecentUsers
List of recent device users, in descending order by last login time (Read-only)

This type is not used in any activity, and only used as part of another schema.

ChromeOsDevices
JSON response template for List Chrome OS Devices operation in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list chromeosdevices (response)

ChromeosdeviceGetMethodBuilder
Retrieve Chrome OS Device

ChromeosdeviceListMethodBuilder
Retrieve all Chrome OS Devices of a customer (paginated)

ChromeosdeviceMethodsBuilder
A builder providing access to all methods supported on chromeosdevice resources. It is not used directly, but through the Directory hub.

ChromeosdevicePatchMethodBuilder
Update Chrome OS Device. This method supports patch semantics.

ChromeosdeviceUpdateMethodBuilder
Update Chrome OS Device

Directory
Central instance to access all Directory related resource activities

Group
JSON template for Group resource in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get groups (response)

aliases insert groups (none)

delete groups (none)

aliases delete groups (none)

patch groups (request|response)

list groups (none)

aliases list groups (none)

update groups (request|response)

insert groups (request|response)

GroupAliaseDeleteMethodBuilder
Remove a alias for the group

GroupAliaseInsertMethodBuilder
Add a alias for the group

GroupAliaseListMethodBuilder
List all aliases for a group

GroupDeleteMethodBuilder
Delete Group

GroupGetMethodBuilder
Retrieve Group

GroupInsertMethodBuilder
Create Group

GroupListMethodBuilder
Retrieve all groups in a domain (paginated)

GroupMethodsBuilder
A builder providing access to all methods supported on group resources. It is not used directly, but through the Directory hub.

GroupPatchMethodBuilder
Update Group. This method supports patch semantics.

GroupUpdateMethodBuilder
Update Group

Groups
JSON response template for List Groups operation in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list groups (response)

Member
JSON template for Member resource in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch members (request|response)

list members (none)

insert members (request|response)

get members (response)

delete members (none)

update members (request|response)

MemberDeleteMethodBuilder
Remove membership.

MemberGetMethodBuilder
Retrieve Group Member

MemberInsertMethodBuilder
Add user to the specified group.

MemberListMethodBuilder
Retrieve all members in a group (paginated)

MemberMethodsBuilder
A builder providing access to all methods supported on member resources. It is not used directly, but through the Directory hub.

MemberPatchMethodBuilder
Update membership of a user in the specified group. This method supports patch semantics.

MemberUpdateMethodBuilder
Update membership of a user in the specified group.

Members
JSON response template for List Members operation in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list members (response)

MobileDevice
JSON template for Mobile Device resource in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get mobiledevices (response)

MobileDeviceAction
JSON request template for firing commands on Mobile Device in Directory Devices API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

action mobiledevices (request)

MobileDeviceApplications
List of applications installed on Mobile Device

This type is not used in any activity, and only used as part of another schema.

MobileDevices
JSON response template for List Mobile Devices operation in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list mobiledevices (response)

MobiledeviceActionMethodBuilder
Take action on Mobile Device

MobiledeviceDeleteMethodBuilder
Delete Mobile Device

MobiledeviceGetMethodBuilder
Retrieve Mobile Device

MobiledeviceListMethodBuilder
Retrieve all Mobile Devices of a customer (paginated)

MobiledeviceMethodsBuilder
A builder providing access to all methods supported on mobiledevice resources. It is not used directly, but through the Directory hub.

Notification
Template for a notification resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete notifications (none)

update notifications (request|response)

patch notifications (request|response)

get notifications (response)

list notifications (none)

NotificationDeleteMethodBuilder
Deletes a notification

NotificationGetMethodBuilder
Retrieves a notification.

NotificationListMethodBuilder
Retrieves a list of notifications.

NotificationMethodsBuilder
A builder providing access to all methods supported on notification resources. It is not used directly, but through the Directory hub.

NotificationPatchMethodBuilder
Updates a notification. This method supports patch semantics.

NotificationUpdateMethodBuilder
Updates a notification.

Notifications
Template for notifications list response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list notifications (response)

OrgUnit
JSON template for Org Unit resource in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch orgunits (request|response)

get orgunits (response)

update orgunits (request|response)

insert orgunits (request|response)

OrgUnits
JSON response template for List Organization Units operation in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list orgunits (response)

OrgunitDeleteMethodBuilder
Remove Organization Unit

OrgunitGetMethodBuilder
Retrieve Organization Unit

OrgunitInsertMethodBuilder
Add Organization Unit

OrgunitListMethodBuilder
Retrieve all Organization Units

OrgunitMethodsBuilder
A builder providing access to all methods supported on orgunit resources. It is not used directly, but through the Directory hub.

OrgunitPatchMethodBuilder
Update Organization Unit. This method supports patch semantics.

OrgunitUpdateMethodBuilder
Update Organization Unit

Schema
JSON template for Schema resource in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get schemas (response)

insert schemas (request|response)

patch schemas (request|response)

delete schemas (none)

list schemas (none)

update schemas (request|response)

SchemaDeleteMethodBuilder
Delete schema

SchemaFieldSpec
JSON template for FieldSpec resource for Schemas in Directory API.

This type is not used in any activity, and only used as part of another schema.

SchemaFieldSpecNumericIndexingSpec
Indexing spec for a numeric field. By default, only exact match queries will be supported for numeric fields. Setting the numericIndexingSpec allows range queries to be supported.

This type is not used in any activity, and only used as part of another schema.

SchemaGetMethodBuilder
Retrieve schema

SchemaInsertMethodBuilder
Create schema.

SchemaListMethodBuilder
Retrieve all schemas for a customer

SchemaMethodsBuilder
A builder providing access to all methods supported on schema resources. It is not used directly, but through the Directory hub.

SchemaPatchMethodBuilder
Update schema. This method supports patch semantics.

SchemaUpdateMethodBuilder
Update schema

Schemas
JSON response template for List Schema operation in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list schemas (response)

Token
JSON template for token resource in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list tokens (none)

delete tokens (none)

get tokens (response)

TokenDeleteMethodBuilder
Delete all access tokens issued by a user for an application.

TokenGetMethodBuilder
Get information about an access token issued by a user.

TokenListMethodBuilder
Returns the set of tokens specified user has issued to 3rd party applications.

TokenMethodsBuilder
A builder providing access to all methods supported on token resources. It is not used directly, but through the Directory hub.

Tokens
JSON response template for List tokens operation in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list tokens (response)

User
JSON template for User object in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

photos patch users (none)

aliases delete users (none)

undelete users (none)

photos get users (none)

update users (request|response)

aliases watch users (none)

insert users (request|response)

photos delete users (none)

patch users (request|response)

photos update users (none)

watch users (none)

get users (response)

aliases insert users (none)

make admin users (none)

aliases list users (none)

list users (none)

delete users (none)

UserAbout
JSON template for About (notes) of a user in Directory API.

This type is not used in any activity, and only used as part of another schema.

UserAddress
JSON template for address.

This type is not used in any activity, and only used as part of another schema.

UserAliaseDeleteMethodBuilder
Remove a alias for the user

UserAliaseInsertMethodBuilder
Add a alias for the user

UserAliaseListMethodBuilder
List all aliases for a user

UserAliaseWatchMethodBuilder
Watch for changes in user aliases list

UserCustomProperties
JSON template for a set of custom properties (i.e. all fields in a particular schema)

This type is not used in any activity, and only used as part of another schema.

UserDeleteMethodBuilder
Delete user

UserEmail
JSON template for an email.

This type is not used in any activity, and only used as part of another schema.

UserExternalId
JSON template for an externalId entry.

This type is not used in any activity, and only used as part of another schema.

UserGetMethodBuilder
retrieve user

UserIm
JSON template for instant messenger of an user.

This type is not used in any activity, and only used as part of another schema.

UserInsertMethodBuilder
create user.

UserListMethodBuilder
Retrieve either deleted users or all users in a domain (paginated)

UserMakeAdmin
JSON request template for setting/revoking admin status of a user in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

make admin users (request)

UserMakeAdminMethodBuilder
change admin status of a user

UserMethodsBuilder
A builder providing access to all methods supported on user resources. It is not used directly, but through the Directory hub.

UserName
JSON template for name of a user in Directory API.

This type is not used in any activity, and only used as part of another schema.

UserOrganization
JSON template for an organization entry.

This type is not used in any activity, and only used as part of another schema.

UserPatchMethodBuilder
update user. This method supports patch semantics.

UserPhone
JSON template for a phone entry.

This type is not used in any activity, and only used as part of another schema.

UserPhoto
JSON template for Photo object in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

photos update users (request|response)

photos patch users (request|response)

photos get users (response)

UserPhotoDeleteMethodBuilder
Remove photos for the user

UserPhotoGetMethodBuilder
Retrieve photo of a user

UserPhotoPatchMethodBuilder
Add a photo for the user. This method supports patch semantics.

UserPhotoUpdateMethodBuilder
Add a photo for the user

UserRelation
JSON template for a relation entry.

This type is not used in any activity, and only used as part of another schema.

UserUndelete
JSON request template to undelete a user in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

undelete users (request)

UserUndeleteMethodBuilder
Undelete a deleted user

UserUpdateMethodBuilder
update user

UserWatchMethodBuilder
Watch for changes in users list

UserWebsite
JSON template for a website entry.

This type is not used in any activity, and only used as part of another schema.

Users
JSON response template for List Users operation in Apps Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list users (response)

VerificationCode
JSON template for verification codes in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

invalidate verification codes (none)

list verification codes (none)

generate verification codes (none)

VerificationCodeGenerateMethodBuilder
Generate new backup verification codes for the user.

VerificationCodeInvalidateMethodBuilder
Invalidate the current backup verification codes for the user.

VerificationCodeListMethodBuilder
Returns the current set of valid backup verification codes for the specified user.

VerificationCodeMethodsBuilder
A builder providing access to all methods supported on verificationCode resources. It is not used directly, but through the Directory hub.

VerificationCodes
JSON response template for List verification codes operation in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list verification codes (response)

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.