Struct google_admin1_directory::ChromeOsDevice

pub struct ChromeOsDevice {
    pub recent_users: Option<Vec<ChromeOsDeviceRecentUsers>>,
    pub mac_address: Option<String>,
    pub serial_number: Option<String>,
    pub last_sync: Option<String>,
    pub order_number: Option<String>,
    pub meid: Option<String>,
    pub annotated_location: Option<String>,
    pub annotated_user: Option<String>,
    pub platform_version: Option<String>,
    pub active_time_ranges: Option<Vec<ChromeOsDeviceActiveTimeRanges>>,
    pub device_id: Option<String>,
    pub ethernet_mac_address: Option<String>,
    pub annotated_asset_id: Option<String>,
    pub firmware_version: Option<String>,
    pub last_enrollment_time: Option<String>,
    pub kind: Option<String>,
    pub os_version: Option<String>,
    pub notes: Option<String>,
    pub will_auto_renew: Option<bool>,
    pub boot_mode: Option<String>,
    pub etag: Option<String>,
    pub status: Option<String>,
    pub org_unit_path: Option<String>,
    pub model: Option<String>,
    pub support_end_date: Option<String>,
}

JSON template for Chrome Os Device resource in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get chromeosdevices (response)

update chromeosdevices (request|response)

patch chromeosdevices (request|response)

Fields

recent_users

List of recent device users, in descending order by last login time (Read-only)

mac_address

Chromebook Mac Address on wifi network interface (Read-only)

serial_number

Chromebook serial number (Read-only)

last_sync

Date and time the device was last synchronized with the policy settings in the Google Apps administrator control panel (Read-only)

order_number

Chromebook order number (Read-only)

meid

Mobile Equipment identifier for the 3G mobile card in the Chromebook (Read-only)

annotated_location

Address or location of the device as noted by the administrator

annotated_user

User of the device

platform_version

Chromebook platform version (Read-only)

active_time_ranges

List of active time ranges (Read-only)

device_id

Unique identifier of Chrome OS Device (Read-only)

ethernet_mac_address

Chromebook Mac Address on ethernet network interface (Read-only)

annotated_asset_id

AssetId specified during enrollment or through later annotation

firmware_version

Chromebook firmware version (Read-only)

last_enrollment_time

Date and time the device was last enrolled (Read-only)

kind

Kind of resource this is.

os_version

Chromebook Os Version (Read-only)

notes

Notes added by the administrator

will_auto_renew

Will Chromebook auto renew after support end date (Read-only)

boot_mode

Chromebook boot mode (Read-only)

etag

ETag of the resource.

status

status of the device (Read-only)

org_unit_path

OrgUnit of the device

model

Chromebook Model (Read-only)

support_end_date

Final date the device will be supported (Read-only)