Struct google_admin1_directory::ChromeOsDevice

pub struct ChromeOsDevice {
    pub recent_users: Option<Vec<ChromeOsDeviceRecentUsers>>,
    pub mac_address: Option<String>,
    pub serial_number: Option<String>,
    pub last_sync: Option<String>,
    pub order_number: Option<String>,
    pub meid: Option<String>,
    pub annotated_location: Option<String>,
    pub annotated_user: Option<String>,
    pub platform_version: Option<String>,
    pub active_time_ranges: Option<Vec<ChromeOsDeviceActiveTimeRanges>>,
    pub system_ram_free_reports: Option<Vec<ChromeOsDeviceSystemRamFreeReports>>,
    pub cpu_status_reports: Option<Vec<ChromeOsDeviceCpuStatusReports>>,
    pub ethernet_mac_address: Option<String>,
    pub system_ram_total: Option<String>,
    pub disk_volume_reports: Option<Vec<ChromeOsDeviceDiskVolumeReports>>,
    pub os_version: Option<String>,
    pub firmware_version: Option<String>,
    pub last_enrollment_time: Option<String>,
    pub kind: Option<String>,
    pub annotated_asset_id: Option<String>,
    pub device_files: Option<Vec<ChromeOsDeviceDeviceFiles>>,
    pub notes: Option<String>,
    pub will_auto_renew: Option<bool>,
    pub boot_mode: Option<String>,
    pub etag: Option<String>,
    pub status: Option<String>,
    pub org_unit_path: Option<String>,
    pub tpm_version_info: Option<ChromeOsDeviceTpmVersionInfo>,
    pub model: Option<String>,
    pub support_end_date: Option<String>,
    pub device_id: Option<String>,
}

JSON template for Chrome Os Device resource in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get chromeosdevices (response)

update chromeosdevices (request|response)

patch chromeosdevices (request|response)

Fields
recent_users: Option<Vec<ChromeOsDeviceRecentUsers>>

List of recent device users, in descending order by last login time (Read-only)

mac_address: Option<String>

Chromebook Mac Address on wifi network interface (Read-only)

serial_number: Option<String>

Chromebook serial number (Read-only)

last_sync: Option<String>

Date and time the device was last synchronized with the policy settings in the G Suite administrator control panel (Read-only)

order_number: Option<String>

Chromebook order number (Read-only)

meid: Option<String>

Mobile Equipment identifier for the 3G mobile card in the Chromebook (Read-only)

annotated_location: Option<String>

Address or location of the device as noted by the administrator

annotated_user: Option<String>

User of the device

platform_version: Option<String>

Chromebook platform version (Read-only)

active_time_ranges: Option<Vec<ChromeOsDeviceActiveTimeRanges>>

List of active time ranges (Read-only)

system_ram_free_reports: Option<Vec<ChromeOsDeviceSystemRamFreeReports>>

Reports of amounts of available RAM memory (Read-only)

cpu_status_reports: Option<Vec<ChromeOsDeviceCpuStatusReports>>

Reports of CPU utilization and temperature (Read-only)

ethernet_mac_address: Option<String>

Chromebook Mac Address on ethernet network interface (Read-only)

system_ram_total: Option<String>

Total RAM on the device [in bytes] (Read-only)

disk_volume_reports: Option<Vec<ChromeOsDeviceDiskVolumeReports>>

Reports of disk space and other info about mounted/connected volumes.

os_version: Option<String>

Chromebook Os Version (Read-only)

firmware_version: Option<String>

Chromebook firmware version (Read-only)

last_enrollment_time: Option<String>

Date and time the device was last enrolled (Read-only)

kind: Option<String>

Kind of resource this is.

annotated_asset_id: Option<String>

AssetId specified during enrollment or through later annotation

device_files: Option<Vec<ChromeOsDeviceDeviceFiles>>

List of device files to download (Read-only)

notes: Option<String>

Notes added by the administrator

will_auto_renew: Option<bool>

Will Chromebook auto renew after support end date (Read-only)

boot_mode: Option<String>

Chromebook boot mode (Read-only)

etag: Option<String>

ETag of the resource.

status: Option<String>

status of the device (Read-only)

org_unit_path: Option<String>

OrgUnit of the device

tpm_version_info: Option<ChromeOsDeviceTpmVersionInfo>

Trusted Platform Module (TPM) (Read-only)

model: Option<String>

Chromebook Model (Read-only)

support_end_date: Option<String>

Final date the device will be supported (Read-only)

device_id: Option<String>

Unique identifier of Chrome OS Device (Read-only)