Struct google_admin1_directory::MobileDevice

pub struct MobileDevice {
    pub last_sync: Option<String>,
    pub device_compromised_status: Option<String>,
    pub resource_id: Option<String>,
    pub default_language: Option<String>,
    pub hardware: Option<String>,
    pub user_agent: Option<String>,
    pub first_sync: Option<String>,
    pub other_accounts_info: Option<Vec<String>>,
    pub release_version: Option<String>,
    pub hardware_id: Option<String>,
    pub etag: Option<String>,
    pub security_patch_level: Option<String>,
    pub privilege: Option<String>,
    pub encryption_status: Option<String>,
    pub type_: Option<String>,
    pub email: Option<Vec<String>>,
    pub device_password_status: Option<String>,
    pub kernel_version: Option<String>,
    pub brand: Option<String>,
    pub meid: Option<String>,
    pub build_number: Option<String>,
    pub applications: Option<Vec<MobileDeviceApplications>>,
    pub device_id: Option<String>,
    pub imei: Option<String>,
    pub adb_status: Option<bool>,
    pub bootloader_version: Option<String>,
    pub manufacturer: Option<String>,
    pub baseband_version: Option<String>,
    pub kind: Option<String>,
    pub developer_options_status: Option<bool>,
    pub name: Option<Vec<String>>,
    pub serial_number: Option<String>,
    pub supports_work_profile: Option<bool>,
    pub network_operator: Option<String>,
    pub status: Option<String>,
    pub unknown_sources_status: Option<bool>,
    pub managed_account_is_on_owner_profile: Option<bool>,
    pub model: Option<String>,
    pub os: Option<String>,
    pub wifi_mac_address: Option<String>,
}

JSON template for Mobile Device resource in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get mobiledevices (response)

Fields
last_sync: Option<String>

Date and time the device was last synchronized with the policy settings in the G Suite administrator control panel (Read-only)

device_compromised_status: Option<String>

Mobile Device compromised status (Read-only)

resource_id: Option<String>

Unique identifier of Mobile Device (Read-only)

default_language: Option<String>

The default locale used on the Mobile Device (Read-only)

hardware: Option<String>

Mobile Device Hardware (Read-only)

user_agent: Option<String>

Mobile Device user agent

first_sync: Option<String>

Date and time the device was first synchronized with the policy settings in the G Suite administrator control panel (Read-only)

other_accounts_info: Option<Vec<String>>

List of accounts added on device (Read-only)

release_version: Option<String>

Mobile Device release version version (Read-only)

hardware_id: Option<String>

Mobile Device Hardware Id (Read-only)

etag: Option<String>

ETag of the resource.

security_patch_level: Option<String>

Mobile Device Security patch level (Read-only)

privilege: Option<String>

DMAgentPermission (Read-only)

encryption_status: Option<String>

Mobile Device Encryption Status (Read-only)

type_: Option<String>

The type of device (Read-only)

email: Option<Vec<String>>

List of owner user's email addresses (Read-only)

device_password_status: Option<String>

DevicePasswordStatus (Read-only)

kernel_version: Option<String>

Mobile Device Kernel version (Read-only)

brand: Option<String>

Mobile Device Brand (Read-only)

meid: Option<String>

Mobile Device MEID number (Read-only)

build_number: Option<String>

Mobile Device Build number (Read-only)

applications: Option<Vec<MobileDeviceApplications>>

List of applications installed on Mobile Device

device_id: Option<String>

Mobile Device serial number (Read-only)

imei: Option<String>

Mobile Device IMEI number (Read-only)

adb_status: Option<bool>

Adb (USB debugging) enabled or disabled on device (Read-only)

bootloader_version: Option<String>

Mobile Device Bootloader version (Read-only)

manufacturer: Option<String>

Mobile Device manufacturer (Read-only)

baseband_version: Option<String>

Mobile Device Baseband version (Read-only)

kind: Option<String>

Kind of resource this is.

developer_options_status: Option<bool>

Developer options enabled or disabled on device (Read-only)

name: Option<Vec<String>>

List of owner user's names (Read-only)

serial_number: Option<String>

Mobile Device SSN or Serial Number (Read-only)

supports_work_profile: Option<bool>

Work profile supported on device (Read-only)

network_operator: Option<String>

Mobile Device mobile or network operator (if available) (Read-only)

status: Option<String>

Status of the device (Read-only)

unknown_sources_status: Option<bool>

Unknown sources enabled or disabled on device (Read-only)

managed_account_is_on_owner_profile: Option<bool>

Boolean indicating if this account is on owner/primary profile or not (Read-only)

model: Option<String>

Name of the model of the device

os: Option<String>

Name of the mobile operating system

wifi_mac_address: Option<String>

Mobile Device WiFi MAC address (Read-only)