Struct google_admin1_directory::MobileDevice

pub struct MobileDevice {
    pub kernel_version: Option<String>,
    pub last_sync: Option<String>,
    pub device_compromised_status: Option<String>,
    pub resource_id: Option<String>,
    pub meid: Option<String>,
    pub default_language: Option<String>,
    pub build_number: Option<String>,
    pub applications: Option<Vec<MobileDeviceApplications>>,
    pub device_id: Option<String>,
    pub imei: Option<String>,
    pub user_agent: Option<String>,
    pub baseband_version: Option<String>,
    pub kind: Option<String>,
    pub name: Option<Vec<String>>,
    pub first_sync: Option<String>,
    pub os: Option<String>,
    pub serial_number: Option<String>,
    pub hardware_id: Option<String>,
    pub network_operator: Option<String>,
    pub etag: Option<String>,
    pub status: Option<String>,
    pub managed_account_is_on_owner_profile: Option<bool>,
    pub model: Option<String>,
    pub type_: Option<String>,
    pub email: Option<Vec<String>>,
    pub wifi_mac_address: Option<String>,
}

JSON template for Mobile Device resource in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get mobiledevices (response)

Fields

kernel_version

Mobile Device Kernel version (Read-only)

last_sync

Date and time the device was last synchronized with the policy settings in the Google Apps administrator control panel (Read-only)

device_compromised_status

Mobile Device compromised status (Read-only)

resource_id

Unique identifier of Mobile Device (Read-only)

meid

Mobile Device MEID number (Read-only)

default_language

The default locale used on the Mobile Device (Read-only)

build_number

Mobile Device Build number (Read-only)

applications

List of applications installed on Mobile Device

device_id

Mobile Device serial number (Read-only)

imei

Mobile Device IMEI number (Read-only)

user_agent

Mobile Device user agent

baseband_version

Mobile Device Baseband version (Read-only)

kind

Kind of resource this is.

name

List of owner user's names (Read-only)

first_sync

Date and time the device was first synchronized with the policy settings in the Google Apps administrator control panel (Read-only)

os

Name of the mobile operating system

serial_number

Mobile Device SSN or Serial Number (Read-only)

hardware_id

Mobile Device Hardware Id (Read-only)

network_operator

Mobile Device mobile or network operator (if available) (Read-only)

etag

ETag of the resource.

status

Status of the device (Read-only)

managed_account_is_on_owner_profile

Boolean indicating if this account is on owner/primary profile or not (Read-only)

model

Name of the model of the device

type_

The type of device (Read-only)

email

List of owner user's email addresses (Read-only)

wifi_mac_address

Mobile Device WiFi MAC address (Read-only)