Struct google_admin1_directory::MobileDevice

pub struct MobileDevice {
    pub kernel_version: Option<String>,
    pub last_sync: Option<String>,
    pub device_compromised_status: Option<String>,
    pub resource_id: Option<String>,
    pub meid: Option<String>,
    pub default_language: Option<String>,
    pub supports_work_profile: Option<bool>,
    pub build_number: Option<String>,
    pub applications: Option<Vec<MobileDeviceApplications>>,
    pub device_id: Option<String>,
    pub imei: Option<String>,
    pub user_agent: Option<String>,
    pub adb_status: Option<bool>,
    pub baseband_version: Option<String>,
    pub kind: Option<String>,
    pub developer_options_status: Option<bool>,
    pub name: Option<Vec<String>>,
    pub first_sync: Option<String>,
    pub other_accounts_info: Option<Vec<String>>,
    pub os: Option<String>,
    pub serial_number: Option<String>,
    pub hardware_id: Option<String>,
    pub network_operator: Option<String>,
    pub etag: Option<String>,
    pub status: Option<String>,
    pub unknown_sources_status: Option<bool>,
    pub managed_account_is_on_owner_profile: Option<bool>,
    pub model: Option<String>,
    pub type_: Option<String>,
    pub email: Option<Vec<String>>,
    pub wifi_mac_address: Option<String>,
}

JSON template for Mobile Device resource in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get mobiledevices (response)

Fields
kernel_version: Option<String>

Mobile Device Kernel version (Read-only)

last_sync: Option<String>

Date and time the device was last synchronized with the policy settings in the Google Apps administrator control panel (Read-only)

device_compromised_status: Option<String>

Mobile Device compromised status (Read-only)

resource_id: Option<String>

Unique identifier of Mobile Device (Read-only)

meid: Option<String>

Mobile Device MEID number (Read-only)

default_language: Option<String>

The default locale used on the Mobile Device (Read-only)

supports_work_profile: Option<bool>

Work profile supported on device (Read-only)

build_number: Option<String>

Mobile Device Build number (Read-only)

applications: Option<Vec<MobileDeviceApplications>>

List of applications installed on Mobile Device

device_id: Option<String>

Mobile Device serial number (Read-only)

imei: Option<String>

Mobile Device IMEI number (Read-only)

user_agent: Option<String>

Mobile Device user agent

adb_status: Option<bool>

Adb (USB debugging) enabled or disabled on device (Read-only)

baseband_version: Option<String>

Mobile Device Baseband version (Read-only)

kind: Option<String>

Kind of resource this is.

developer_options_status: Option<bool>

Developer options enabled or disabled on device (Read-only)

name: Option<Vec<String>>

List of owner user's names (Read-only)

first_sync: Option<String>

Date and time the device was first synchronized with the policy settings in the Google Apps administrator control panel (Read-only)

other_accounts_info: Option<Vec<String>>

List of accounts added on device (Read-only)

os: Option<String>

Name of the mobile operating system

serial_number: Option<String>

Mobile Device SSN or Serial Number (Read-only)

hardware_id: Option<String>

Mobile Device Hardware Id (Read-only)

network_operator: Option<String>

Mobile Device mobile or network operator (if available) (Read-only)

etag: Option<String>

ETag of the resource.

status: Option<String>

Status of the device (Read-only)

unknown_sources_status: Option<bool>

Unknown sources enabled or disabled on device (Read-only)

managed_account_is_on_owner_profile: Option<bool>

Boolean indicating if this account is on owner/primary profile or not (Read-only)

model: Option<String>

Name of the model of the device

type_: Option<String>

The type of device (Read-only)

email: Option<Vec<String>>

List of owner user's email addresses (Read-only)

wifi_mac_address: Option<String>

Mobile Device WiFi MAC address (Read-only)