Struct google_admin1_directory::SchemaFieldSpec

pub struct SchemaFieldSpec {
    pub kind: Option<String>,
    pub etag: Option<String>,
    pub read_access_type: Option<String>,
    pub field_type: Option<String>,
    pub indexed: Option<bool>,
    pub field_name: Option<String>,
    pub numeric_indexing_spec: Option<SchemaFieldSpecNumericIndexingSpec>,
    pub field_id: Option<String>,
    pub multi_valued: Option<bool>,
}

JSON template for FieldSpec resource for Schemas in Directory API.

This type is not used in any activity, and only used as part of another schema.

Fields

kind

Kind of resource this is.

etag

ETag of the resource.

read_access_type

Read ACLs on the field specifying who can view values of this field. Valid values are "ALL_DOMAIN_USERS" and "ADMINS_AND_SELF".

field_type

Type of the field.

indexed

Boolean specifying whether the field is indexed or not.

field_name

Name of the field.

numeric_indexing_spec

Indexing spec for a numeric field. By default, only exact match queries will be supported for numeric fields. Setting the numericIndexingSpec allows range queries to be supported.

field_id

Unique identifier of Field (Read-only)

multi_valued

Boolean specifying whether this is a multi-valued field or not.