Struct google_admin1_directory::User

pub struct User {
    pub addresses: Option<String>,
    pub phones: Option<String>,
    pub is_delegated_admin: Option<bool>,
    pub thumbnail_photo_etag: Option<String>,
    pub suspended: Option<bool>,
    pub id: Option<String>,
    pub aliases: Option<Vec<String>>,
    pub non_editable_aliases: Option<Vec<String>>,
    pub custom_schemas: Option<HashMap<String, UserCustomProperties>>,
    pub deletion_time: Option<String>,
    pub suspension_reason: Option<String>,
    pub thumbnail_photo_url: Option<String>,
    pub is_admin: Option<bool>,
    pub relations: Option<String>,
    pub include_in_global_address_list: Option<bool>,
    pub etag: Option<String>,
    pub last_login_time: Option<String>,
    pub org_unit_path: Option<String>,
    pub agreed_to_terms: Option<bool>,
    pub external_ids: Option<String>,
    pub ip_whitelisted: Option<bool>,
    pub kind: Option<String>,
    pub is_mailbox_setup: Option<bool>,
    pub password: Option<String>,
    pub emails: Option<String>,
    pub organizations: Option<String>,
    pub primary_email: Option<String>,
    pub hash_function: Option<String>,
    pub name: Option<UserName>,
    pub notes: Option<String>,
    pub creation_time: Option<String>,
    pub websites: Option<String>,
    pub change_password_at_next_login: Option<bool>,
    pub ims: Option<String>,
    pub customer_id: Option<String>,
}

JSON template for User object in Directory API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

photos patch users (none)

aliases delete users (none)

undelete users (none)

photos get users (none)

update users (request|response)

aliases watch users (none)

insert users (request|response)

photos delete users (none)

patch users (request|response)

photos update users (none)

watch users (none)

get users (response)

aliases insert users (none)

make admin users (none)

aliases list users (none)

list users (none)

delete users (none)

Fields

addresses

no description provided

phones

no description provided

is_delegated_admin

Boolean indicating if the user is delegated admin (Read-only)

thumbnail_photo_etag

ETag of the user's photo (Read-only)

suspended

Indicates if user is suspended

id

Unique identifier of User (Read-only)

aliases

List of aliases (Read-only)

non_editable_aliases

List of non editable aliases (Read-only)

custom_schemas

Custom fields of the user.

deletion_time

no description provided

suspension_reason

Suspension reason if user is suspended (Read-only)

thumbnail_photo_url

Photo Url of the user (Read-only)

is_admin

Boolean indicating if the user is admin (Read-only)

relations

no description provided

include_in_global_address_list

Boolean indicating if user is included in Global Address List

etag

ETag of the resource.

last_login_time

User's last login time. (Read-only)

org_unit_path

OrgUnit of User

agreed_to_terms

Indicates if user has agreed to terms (Read-only)

external_ids

no description provided

ip_whitelisted

Boolean indicating if ip is whitelisted

kind

Kind of resource this is.

is_mailbox_setup

Is mailbox setup (Read-only)

password

User's password

emails

no description provided

organizations

no description provided

primary_email

username of User

hash_function

Hash function name for password. Supported are MD5, SHA-1 and crypt

name

User's name

notes

no description provided

creation_time

User's Google account creation time. (Read-only)

websites

no description provided

change_password_at_next_login

Boolean indicating if the user should change password in next login

ims

no description provided

customer_id

CustomerId of User (Read-only)