Struct google_admin1_directory::UserMethods

pub struct UserMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on user resources. It is not used directly, but through the Directory hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_admin1_directory as admin1_directory;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use admin1_directory::Directory;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Directory::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `aliases_delete(...)`, `aliases_insert(...)`, `aliases_list(...)`, `aliases_watch(...)`, `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `make_admin(...)`, `patch(...)`, `photos_delete(...)`, `photos_get(...)`, `photos_patch(...)`, `photos_update(...)`, `undelete(...)`, `update(...)` and `watch(...)`
// to build up your call.
let rb = hub.users();

Methods

impl<'a, C, A> UserMethods<'a, C, A>

fn undelete(&self, request: UserUndelete, user_key: &str) -> UserUndeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Undelete a deleted user

Arguments

request - No description provided.

userKey - The immutable id of the user

fn photos_delete(&self, user_key: &str) -> UserPhotoDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Remove photos for the user

Arguments

userKey - Email or immutable Id of the user

fn photos_patch(&self, request: UserPhoto, user_key: &str) -> UserPhotoPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Add a photo for the user. This method supports patch semantics.

Arguments

request - No description provided.

userKey - Email or immutable Id of the user

fn insert(&self, request: User) -> UserInsertCall<'a, C, A>

Create a builder to help you perform the following task:

create user.

Arguments

request - No description provided.

fn aliases_watch(&self, request: Channel, user_key: &str) -> UserAliaseWatchCall<'a, C, A>

Create a builder to help you perform the following task:

Watch for changes in user aliases list

Arguments

request - No description provided.

userKey - Email or immutable Id of the user

fn photos_update(&self, request: UserPhoto, user_key: &str) -> UserPhotoUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Add a photo for the user

Arguments

request - No description provided.

userKey - Email or immutable Id of the user

fn watch(&self, request: Channel) -> UserWatchCall<'a, C, A>

Create a builder to help you perform the following task:

Watch for changes in users list

Arguments

request - No description provided.

fn update(&self, request: User, user_key: &str) -> UserUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

update user

Arguments

request - No description provided.

userKey - Email or immutable Id of the user. If Id, it should match with id of user object

fn photos_get(&self, user_key: &str) -> UserPhotoGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieve photo of a user

Arguments

userKey - Email or immutable Id of the user

fn list(&self) -> UserListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieve either deleted users or all users in a domain (paginated)

fn get(&self, user_key: &str) -> UserGetCall<'a, C, A>

Create a builder to help you perform the following task:

retrieve user

Arguments

userKey - Email or immutable Id of the user

fn aliases_insert(&self, request: Alias, user_key: &str) -> UserAliaseInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Add a alias for the user

Arguments

request - No description provided.

userKey - Email or immutable Id of the user

fn aliases_list(&self, user_key: &str) -> UserAliaseListCall<'a, C, A>

Create a builder to help you perform the following task:

List all aliases for a user

Arguments

userKey - Email or immutable Id of the user

fn make_admin(&self, request: UserMakeAdmin, user_key: &str) -> UserMakeAdminCall<'a, C, A>

Create a builder to help you perform the following task:

change admin status of a user

Arguments

request - No description provided.

userKey - Email or immutable Id of the user as admin

fn patch(&self, request: User, user_key: &str) -> UserPatchCall<'a, C, A>

Create a builder to help you perform the following task:

update user. This method supports patch semantics.

Arguments

request - No description provided.

userKey - Email or immutable Id of the user. If Id, it should match with id of user object

fn aliases_delete(&self, user_key: &str, alias: &str) -> UserAliaseDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Remove a alias for the user

Arguments

userKey - Email or immutable Id of the user

alias - The alias to be removed

fn delete(&self, user_key: &str) -> UserDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete user

Arguments

userKey - Email or immutable Id of the user